refactor: Flatten dist folder
