+++ lib/index.ts
@@ -1,3 +1,3 @@
-export * from './links';
+export * from './headersLink';
 export * from './types';
 export * from './useRestQuery';
